General results caching infrastructure update
Introducing a GENERAL_RESULTS_CACHE array.
The keys will be:
  event_results: (this the same as the EVENT_RESULTS_CACHE that was originally there.

 race_list:  This caches the race_list results (this is used for marshalling and also for pilot results viewing

classX (where X is the class id): this holds the class summary results.

Each key has 2 elements:
  isvalid and  content..
isvalid=0 when the content is stale
isvalid=1 when the content is good
content is the cached data.

We may be able to utilize this in the future. when generating the overall report to cache sub sections of the report, and then use the sections to build the overall event report..