pilot sorting, heat generator, pilot attributes, display heat/round of fastest lap
This change adds the following functions:
Pilots can be sorted by name or callsign. This is reflected both on the pilots screen, and in the pulldowns when assigning pilots in heats.

Heat generator based on results of a previous runs.   This generates letter type mains (a-main, b-main, c-main) heats based on selecting win condtion of  either most laps, fastest 3 consecutive, or fastest lap.

Pilot Attributes:  Race directors can now define attributes for a pilot such as 'Registered' , or passed tech, paid to help with bookeeping,  The attribute named 'Registered' is special as only 'Registered' pilots will appear in the pilots list of heat pulldowns.

To 'un-Register' all pilots, specify Registered in text field and press the Delete Attribute button.

 The results page Fastest lap now indictes which heat and which round that pilot achieved their fastest lap.  This is most helpful when looking over results to determine if a specific heat/round/pilot should be marshalled because of an unusually fast lap.