results page update for class and pilot summary results
The existing results page calculates ALL results, which takes over 1 minute to render.  While the async code does allow the PI to continue working it does affect performance, and also it is very long for a pilot to see his/her results after a race..

This commit addresses these issues:
a. The results page does not automatically fetch all of the rounds data,  instead it allows the user to select which type of report they want.
 - Spectators (non-admins) can select either  results by pilot:  in which case they can narrow down to the specific heat, round, pilot to see every lap time.  This looks very similar to the marshalling results without the graph, and  no ability to change anything from this page, only view.

results by class:  in which case  they select which class that they want to see the summary  results.  This is most often used when running fastest lap, or fastest 3 consecutive for a qualifying class, the pilot can see those results quickly after each race.

The admin user can also select the Total Event Results, which is the standard get everything.
I feel that limiting this to only admins is proper and really minimizes both load on the PI and network traffic generated by many pilots looking for results.   However, it is desired to allow any pilot to select this, simply remove the admin-hide tags found in the rounds.html

The 'EVENT_RESULTS_CACHE' has been replaced with a GENERAL_RESULTS_CACHE.  The GENERAL_RESULTS_CACHE allows for both the total event_results to be cached, and also each individual class results as well.

I've used this concept for 5+ races over the last 3 months and the pilots had no complaints. They were able to see the summary that they wanted, and the pilot details with minimal impact to PI processing.