Cleanup VRxNode classes to support broadcast
Added in base VRxNode class to inherit from.
Added basic support and bugfix for the broadcast node.
Fixed set_message for all units
Added in Event register for when admin Sends Message to show up on CV OSD.
Added VRxController startup behavior to disable parts of the OSD
Fixed bug with trying to set frequencies that didn't exist in the CV but did in the python API
Added support for race finish message