for some reason, find_line() isn't working when using chip iterator, so iI had to loop over the lines to find the one with the matching name. and of course line.name is not a string, it is a property so you have to get it before you can compare it. just doing the_line.name == 'string' wont work because it compares a string to a property instead of string to string.
Added a tool to manually test the button press listener class without running the whole server.

Currently we have to manually input the pin as pin number on potato, but as gpio number on raspberry pi. I have not figured out a good way to map that yet.
might just have to build a static mapping table.