Add a way to set correct node to pilot on first heat
During races, we do not only have R-band pilots.
Most of the time, few DJI/O3 pilots are also there.

So far, managing such pilots was painful as it needed a manual change
either in nodes settings or heat pilot seats in order to have correct
DJI pilots configured on correct nodes.

With this new plugin and associated patches, DJI pilots are
automatically assigned to a D or O band node if available.

This is mostly useful on first heat, when no history has been recorded
yet. On subsequent heats, the current algorithm would reuse the previous
frequencies.

This commit is also introducing a new configuration for 8 nodes RH with:
- R1367 + D67 + O67

D67 and O67 are configured for DJI/DJI O3 pilots.

Note that:
- when D6 or O6 is select, R7 and O6 or D6 are then discarded
- when D7 or O7 is select, O7 or D7 are then discarded.

Fix bug GH-468

Signed-off-by: Arnaud M <arnaud.choupinette@mailops.fr>